moved CONSTANTS to server side
